# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Auth Service.",
        "title": "Auth Service API3",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4444",
    "paths": {
        "/center/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint yangi qayta ishlash markazini yaratish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "center"
                ],
                "summary": "Yangi qayta ishlash markazi yaratish",
                "parameters": [
                    {
                        "description": "Recycling Center Payload",
                        "name": "recycling_center",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recycling_center.ResCenter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "$ref": "#/definitions/recycling_center.ResponceResCenter"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/center/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint qayta ishlash markazini qidirish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "center"
                ],
                "summary": "Qayta ishlash markazini qidirish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Material type",
                        "name": "material",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "$ref": "#/definitions/recycling_center.ResAllCenter"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/center/statistics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint statistika ma'lumotlarini olish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "center"
                ],
                "summary": "Statistika olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date in format yyyy-mm-dd",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date in format yyyy-mm-dd",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "schema": {
                            "$ref": "#/definitions/recycling_center.Statistics"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "recycling_center.Center": {
            "type": "object",
            "properties": {
                "acceptedMaterials": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "address": {
                    "type": "string"
                },
                "contactNumber": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "workingHours": {
                    "type": "string"
                }
            }
        },
        "recycling_center.ResAllCenter": {
            "type": "object",
            "properties": {
                "centers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/recycling_center.Center"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "recycling_center.ResCenter": {
            "type": "object",
            "properties": {
                "acceptedMaterials": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "address": {
                    "type": "string"
                },
                "contactNumber": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "workingHours": {
                    "type": "string"
                }
            }
        },
        "recycling_center.ResponceResCenter": {
            "type": "object",
            "properties": {
                "acceptedMaterials": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "address": {
                    "type": "string"
                },
                "contactNumber": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "workingHours": {
                    "type": "string"
                }
            }
        },
        "recycling_center.Statistics": {
            "type": "object",
            "properties": {
                "TotalEarnedPoints": {
                    "type": "integer"
                },
                "TotalRecycledItems": {
                    "type": "integer"
                },
                "TotalSwaps": {
                    "type": "integer"
                },
                "jsonDataTopCateg": {
                    "type": "string"
                },
                "jsonDateTopCenter": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}